refactor: 운영진용 API 분리
