+++ src/main/java/com/umc/networkingService/domain/project/controller/ProjectController.java
@@ -26,42 +26,6 @@ import java.util.UUID;
 public class ProjectController {
     private final ProjectService projectService;
 
-    @Operation(summary = "프로젝트 등록 API", description = "프로젝트를 등록하는 API입니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "COMMON200", description = "성공")
-    })
-    @PostMapping
-    public BaseResponse<ProjectIdResponse> createProject(@CurrentMember Member member,
-                                                         @RequestPart MultipartFile projectImage,
-                                                         @Valid @RequestPart ProjectCreateRequest request){
-        return BaseResponse.onSuccess(projectService.createProject(member, projectImage, request));
-    }
-
-    @Operation(summary = "프로젝트 수정 API", description = "프로젝트를 수정하는 API입니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "COMMON200", description = "성공"),
-            @ApiResponse(responseCode = "PROJECT001", description = "존재하지 않는 프로젝트 입니다."),
-            @ApiResponse(responseCode = "PROJECT002", description = "해당 프로젝트에 대해 수정 권한이 없습니다.")
-    })
-    @PostMapping("/update/{projectId}")
-    public BaseResponse<ProjectIdResponse> updateProject(@CurrentMember Member member,
-                                                             @PathVariable ("projectId") UUID projectId,
-                                                             @RequestPart MultipartFile projectImage,
-                                                             @Valid @RequestPart ProjectUpdateRequest request){
-        return BaseResponse.onSuccess(projectService.updateProject(member, projectId, projectImage, request));
-    }
-
-    @Operation(summary = "프로젝트 삭제 API", description = "프로젝트를 삭제하는 API입니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "COMMON200", description = "성공"),
-            @ApiResponse(responseCode = "PROJECT001", description = "존재하지 않는 프로젝트 입니다."),
-            @ApiResponse(responseCode = "PROJECT003", description = "해당 프로젝트에 대해 삭제 권한이 없습니다.")
-    })
-    @DeleteMapping("/{projectId}")
-    public BaseResponse<ProjectIdResponse> deleteProject(@PathVariable("projectId") UUID projectId){
-        return BaseResponse.onSuccess(projectService.deleteProject(projectId));
-    }
-
     @GetMapping
     @Operation(summary = "프로젝트 조회 API", description = "프로젝트를 조회하는 API입니다.")
     @ApiResponses(value = {
